requirements-dev.txt
# Development dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
pytest-mock==3.12.0
flake8==6.1.0
black==23.11.0